# src/main/java/com/josue/kodeur/xtremanalyse/api/api-docs.json
{
  "openapi": "3.0.1" ,
  "info": {
    "title": "OpenAPI definition, Xtrem-Analyse" ,
    "version": "v0"
  } ,
  "servers": [
    {
      "url": "http://localhost:9000" ,
      "description": "Generated server url"
    }
  ] ,
  "paths": {
...
    "/api/v1/user/update": {
      "put": {
        "tags": [
          "user-controller"
        ] ,
        "operationId": "updateUser" ,
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddForm"
              }
            }
          } ,
          "required": true
        } ,
        "responses": {
          "400": {
            "description": "Bad Request" ,
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponse"
                }
              }
            }
          } ,
          "401": {
            "description": "Unauthorized" ,
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponse"
                }
              }
            }
          } ,
          "200": {
            "description": "OK" ,
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        }
      }
    } ,
    "/api/v1/user/delete-role": {
      "put": {
        "tags": [
          "user-controller"
        ] ,
        "operationId": "deleteRole" ,
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteRoleForm"
              }
            }
          } ,
          "required": true
        } ,
        "responses": {
          "400": {
            "description": "Bad Request" ,
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponse"
                }
              }
            }
          } ,
          "401": {
            "description": "Unauthorized" ,
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponse"
                }
              }
            }
          } ,
          "200": {
            "description": "OK"
          }
        }
      }
    } ,
...
    "/api/v1/user/register": {
      "post": {
        "tags": [
          "user-controller"
        ] ,
        "operationId": "register" ,
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddForm"
              }
            }
          } ,
          "required": true
        } ,
        "responses": {
          "400": {
            "description": "Bad Request" ,
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponse"
                }
              }
            }
          } ,
          "401": {
            "description": "Unauthorized" ,
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponse"
                }
              }
            }
          } ,
          "200": {
            "description": "OK" ,
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        }
      }
    } ,
    "/api/v1/user/login": {
      "post": {
        "tags": [
          "user-controller"
        ] ,
        "operationId": "login" ,
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginForm"
              }
            }
          } ,
          "required": true
        } ,
        "responses": {
          "400": {
            "description": "Bad Request" ,
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponse"
                }
              }
            }
          } ,
          "401": {
            "description": "Unauthorized" ,
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponse"
                }
              }
            }
          } ,
          "200": {
            "description": "OK" ,
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        }
      }
    } ,
    "/api/v1/user/add": {
      "post": {
        "tags": [
          "user-controller"
        ] ,
        "operationId": "add_1" ,
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddForm"
              }
            }
          } ,
          "required": true
        } ,
        "responses": {
          "400": {
            "description": "Bad Request" ,
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponse"
                }
              }
            }
          } ,
          "401": {
            "description": "Unauthorized" ,
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponse"
                }
              }
            }
          } ,
          "200": {
            "description": "OK" ,
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        }
      }
    } ,
...
    "/api/v1/user/search/{matricule}": {
      "get": {
        "tags": [
          "user-controller"
        ] ,
        "operationId": "search" ,
        "parameters": [
          {
            "name": "matricule" ,
            "in": "path" ,
            "required": true ,
            "schema": {
              "type": "string"
            }
          }
        ] ,
        "responses": {
          "400": {
            "description": "Bad Request" ,
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponse"
                }
              }
            }
          } ,
          "401": {
            "description": "Unauthorized" ,
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponse"
                }
              }
            }
          } ,
          "200": {
            "description": "OK" ,
            "content": {
              "*/*": {
                "schema": {
                  "type": "array" ,
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          }
        }
      }
    } ,
    "/api/v1/user/list-users": {
      "get": {
        "tags": [
          "user-controller"
        ] ,
        "operationId": "allUsers" ,
        "responses": {
          "400": {
            "description": "Bad Request" ,
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponse"
                }
              }
            }
          } ,
          "401": {
            "description": "Unauthorized" ,
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponse"
                }
              }
            }
          } ,
          "200": {
            "description": "OK" ,
            "content": {
              "*/*": {
                "schema": {
                  "type": "array" ,
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          }
        }
      }
    } ,
    "/api/v1/user/find/{matricule}": {
      "get": {
        "tags": [
          "user-controller"
        ] ,
        "operationId": "profile" ,
        "parameters": [
          {
            "name": "matricule" ,
            "in": "path" ,
            "required": true ,
            "schema": {
              "type": "string"
            }
          }
        ] ,
        "responses": {
          "400": {
            "description": "Bad Request" ,
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponse"
                }
              }
            }
          } ,
          "401": {
            "description": "Unauthorized" ,
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponse"
                }
              }
            }
          } ,
          "200": {
            "description": "OK" ,
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        }
      }
    } ,
...
    "/api/v1/reports/userList/": {
      "get": {
        "tags": [
          "report-controller"
        ] ,
        "operationId": "userList" ,
        "responses": {
          "400": {
            "description": "Bad Request" ,
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponse"
                }
              }
            }
          } ,
          "401": {
            "description": "Unauthorized" ,
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponse"
                }
              }
            }
          } ,
          "200": {
            "description": "OK" ,
            "content": {
              "*/*": {
                "schema": {
                  "type": "string" ,
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    } ,
...
    "/api/v1/user/delete/{id}": {
      "delete": {
        "tags": [
          "user-controller"
        ] ,
        "operationId": "deleteUser" ,
        "parameters": [
          {
            "name": "id" ,
            "in": "path" ,
            "required": true ,
            "schema": {
              "type": "integer" ,
              "format": "int64"
            }
          }
        ] ,
        "responses": {
          "400": {
            "description": "Bad Request" ,
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponse"
                }
              }
            }
          } ,
          "401": {
            "description": "Unauthorized" ,
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponse"
                }
              }
            }
          } ,
          "200": {
            "description": "OK"
          }
        }
      }
    } ,
...
  } ,
  "components": {
    "schemas": {
      "HttpResponse": {
        "type": "object" ,
        "properties": {
          "timestamp": {
            "type": "string" ,
            "format": "date-time"
          } ,
          "statusCode": {
            "type": "integer" ,
            "format": "int32"
          } ,
          "httpStatus": {
            "type": "string" ,
            "enum": [
              "100 CONTINUE" ,
              "101 SWITCHING_PROTOCOLS" ,
              "102 PROCESSING" ,
              "103 CHECKPOINT" ,
              "200 OK" ,
              "201 CREATED" ,
              "202 ACCEPTED" ,
              "203 NON_AUTHORITATIVE_INFORMATION" ,
              "204 NO_CONTENT" ,
              "205 RESET_CONTENT" ,
              "206 PARTIAL_CONTENT" ,
              "207 MULTI_STATUS" ,
              "208 ALREADY_REPORTED" ,
              "226 IM_USED" ,
              "300 MULTIPLE_CHOICES" ,
              "301 MOVED_PERMANENTLY" ,
              "302 FOUND" ,
              "302 MOVED_TEMPORARILY" ,
              "303 SEE_OTHER" ,
              "304 NOT_MODIFIED" ,
              "305 USE_PROXY" ,
              "307 TEMPORARY_REDIRECT" ,
              "308 PERMANENT_REDIRECT" ,
              "400 BAD_REQUEST" ,
              "401 UNAUTHORIZED" ,
              "402 PAYMENT_REQUIRED" ,
              "403 FORBIDDEN" ,
              "404 NOT_FOUND" ,
              "405 METHOD_NOT_ALLOWED" ,
              "406 NOT_ACCEPTABLE" ,
              "407 PROXY_AUTHENTICATION_REQUIRED" ,
              "408 REQUEST_TIMEOUT" ,
              "409 CONFLICT" ,
              "410 GONE" ,
              "411 LENGTH_REQUIRED" ,
              "412 PRECONDITION_FAILED" ,
              "413 PAYLOAD_TOO_LARGE" ,
              "413 REQUEST_ENTITY_TOO_LARGE" ,
              "414 URI_TOO_LONG" ,
              "414 REQUEST_URI_TOO_LONG" ,
              "415 UNSUPPORTED_MEDIA_TYPE" ,
              "416 REQUESTED_RANGE_NOT_SATISFIABLE" ,
              "417 EXPECTATION_FAILED" ,
              "418 I_AM_A_TEAPOT" ,
              "419 INSUFFICIENT_SPACE_ON_RESOURCE" ,
              "420 METHOD_FAILURE" ,
              "421 DESTINATION_LOCKED" ,
              "422 UNPROCESSABLE_ENTITY" ,
              "423 LOCKED" ,
              "424 FAILED_DEPENDENCY" ,
              "425 TOO_EARLY" ,
              "426 UPGRADE_REQUIRED" ,
              "428 PRECONDITION_REQUIRED" ,
              "429 TOO_MANY_REQUESTS" ,
              "431 REQUEST_HEADER_FIELDS_TOO_LARGE" ,
              "451 UNAVAILABLE_FOR_LEGAL_REASONS" ,
              "500 INTERNAL_SERVER_ERROR" ,
              "501 NOT_IMPLEMENTED" ,
              "502 BAD_GATEWAY" ,
              "503 SERVICE_UNAVAILABLE" ,
              "504 GATEWAY_TIMEOUT" ,
              "505 HTTP_VERSION_NOT_SUPPORTED" ,
              "506 VARIANT_ALSO_NEGOTIATES" ,
              "507 INSUFFICIENT_STORAGE" ,
              "508 LOOP_DETECTED" ,
              "509 BANDWIDTH_LIMIT_EXCEEDED" ,
              "510 NOT_EXTENDED" ,
              "511 NETWORK_AUTHENTICATION_REQUIRED"
            ]
          } ,
          "reason": {
            "type": "string"
          } ,
          "message": {
            "type": "string"
          }
        }
      } ,
...
      "Role": {
        "type": "object" ,
        "properties": {
          "id": {
            "type": "integer" ,
            "format": "int64"
          } ,
          "nom": {
            "type": "string"
          }
        }
      } ,
...
      "User": {
        "type": "object" ,
        "properties": {
          "id": {
            "type": "integer" ,
            "format": "int64"
          } ,
          "userMatricule": {
            "type": "string"
          } ,
          "nom": {
            "type": "string"
          } ,
          "password": {
            "type": "string" ,
            "writeOnly": true
          } ,
          "lastLoginDate": {
            "type": "string" ,
            "format": "date-time"
          } ,
          "lastLoginDateDisplay": {
            "type": "string" ,
            "format": "date-time"
          } ,
          "joinDate": {
            "type": "string" ,
            "format": "date-time"
          } ,
          "isActive": {
            "type": "boolean"
          } ,
          "isNotLocked": {
            "type": "boolean"
          } ,
          "token": {
            "type": "string"
          } ,
          "roles": {
            "type": "array" ,
            "items": {
              "$ref": "#/components/schemas/Role"
            }
          } ,
          "tokenValid": {
            "type": "boolean"
          }
        }
      } ,
...
      "AddForm": {
        "type": "object" ,
        "properties": {
          "matricule": {
            "type": "string"
          } ,
          "newMatricule": {
            "type": "string"
          } ,
          "password": {
            "type": "string"
          } ,
          "nom": {
            "type": "string"
          } ,
          "isActive": {
            "type": "string"
          } ,
          "isNotLocked": {
            "type": "string"
          } ,
          "role": {
            "type": "string"
          }
        }
      } ,
      "DeleteRoleForm": {
        "type": "object" ,
        "properties": {
          "matricule": {
            "type": "string"
          } ,
          "roles": {
            "type": "array" ,
            "items": {
              "type": "string"
            }
          }
        }
      } ,
...
      "LoginForm": {
        "type": "object" ,
        "properties": {
          "matricule": {
            "type": "string"
          } ,
          "password": {
            "type": "string"
          }
        }
      } ,
...
    }
  }
}